pyproject.toml
[tool.poetry]
name = "pytorch-env"
version = "0.1.0"
description = ""
authors = ["Alejandro Amar <alejandroamar66@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = {version="^2.2.1", source="torch"}
torchaudio = { version = "^2.2.0", source="torch"}
torchvision = { version = "^0.17", source="torch"}
matplotlib = "^3.8.3"
scikit-learn = "^1.4.1.post1"
scikit-image = "^0.22.0"
opencv-python = "^4.9.0.80"
yacs = "^0.1.8"
joblib = "^1.3.2"
natsort = "^8.4.0"
h5py = "^3.10.0"
tqdm = "^4.66.2"
einops = "^0.7.0"
gdown = "^5.1.0"
addict = "^2.4.0"
future = "^1.0.0"
lmdb = "^1.4.1"
numpy = "^1.26.4"
pyyaml = "^6.0.1"
requests = "^2.31.0"
scipy = "^1.12.0"
tb-nightly = "^2.17.0a20240318"
yapf = "^0.40.2"
lpips = "^0.1.4"
icecream = "^2.1.3"
mypy = "^1.9.0"
pandas = "^2.2.2"
prettytable = "^3.10.0"
torchsummary = "^1.5.1"
linformer = "^0.2.3"
timm = "^0.9.16"
ptflops = "^0.7.2.2"
tornado = "^6.4"
jupyterlab = "^4.2.0"

[[tool.poetry.source]]

name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority  = "supplemental"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
